rename "api" to "core" package
