+++ src/core/configuration.go
@@ -1,4 +1,4 @@
-package api
+package core
 
 import (
 	"StationeersServerUI/src/config"
